finished all of the functions of chat

+++ src/main/java/com/tudo/chatify/repository/UserRepository.java
@@ -10,8 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-//        User findByUsername(String username);
+public interface UserRepository extends JpaRepository<User, Integer> {
         User findByEmail(String email) throws UserException;
         @Query("SELECT u FROM User u WHERE u.full_name Like %:query% OR u.email Like %:query%")
         public List<User> searchUser(@Param("Query") String query);
